requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
cfgv==3.4.0
decorator==5.1.1
distlib==0.3.7
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==1.2.0
filelock==3.12.4
identify==2.5.29
ipython==8.15.0
jedi==0.19.0
Markdown==3.4.4
matplotlib-inline==0.1.6
nodeenv==1.8.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.0
virtualenv==20.24.5
wcwidth==0.2.6
